backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai
pydub==0.25.1
pymupdf==1.25.1
fpdf2==2.8.2
python-dotenv==1.0.1
google-genai
audioop-lts
azure-cognitiveservices-speech
moviepy
